+++ src/codeoceanauxtools/get_cpu_count.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
-import util
+import co_aux_utils
 
 
 def main():
-    return util.get_cpu_limit()
+    return co_aux_utils.get_cpu_limit()
 
 
 if __name__ == "__main__":
