+++ src/codeoceanauxtools/get_dir_contents.py
@@ -2,18 +2,18 @@
 import sys
 
 from get_logger import LOGGER
-import util
+import co_aux_utils
 
 
 def main(argv=None):
     if len(argv) > 1:
         return (
             f"*** These are the current files in the {argv[1]} directory\n"
-            + f"{util.get_dir_contents(argv[1])}"
+            + f"{co_aux_utils.get_dir_contents(argv[1])}"
         )
     return (
         "*** These are the current files in the ../data directory\n"
-        + f"{util.get_dir_contents()}"
+        + f"{co_aux_utils.get_dir_contents()}"
     )
