+++ src/codeoceanauxtools/get_fwd_fastqs.py
@@ -2,14 +2,14 @@
 import sys
 
 from get_logger import LOGGER
-import util
+import co_aux_utils
 
 
 def main(argv=None):
     if len(argv) > 1:
         LOGGER.info(f"Searching in {argv[1]} dir for files")
-        return util.get_fwd_fastqs(argv[1])
-    return util.get_fwd_fastqs()
+        return co_aux_utils.get_fwd_fastqs(argv[1])
+    return co_aux_utils.get_fwd_fastqs()
 
 
 if __name__ == "__main__":
