+++ src/codeoceanauxtools/get_groups.py
@@ -2,13 +2,13 @@
 import sys
 
 from get_logger import LOGGER
-import util
+import co_aux_utils
 
 
 def main(argv=None):
     if len(argv) > 1:
-        return util.get_groups(argv[1])
-    return util.get_groups()
+        return co_aux_utils.get_groups(argv[1])
+    return co_aux_utils.get_groups()
 
 
 if __name__ == "__main__":
