+++ src/codeoceanauxtools/get_read_pattern.py
@@ -2,7 +2,7 @@
 import sys
 
 from get_logger import LOGGER
-import util
+import co_aux_utils
 
 
 def main(argv=None):
@@ -10,8 +10,8 @@ def main(argv=None):
         LOGGER.error("You failed to provide a filename and direction")
         sys.exit("You failed to provide a filename and direction")
     elif len(argv) == 2:
-        return util.get_read_pattern(argv[1])
-    return util.get_read_pattern(argv[1], argv[2])
+        return co_aux_utils.get_read_pattern(argv[1])
+    return co_aux_utils.get_read_pattern(argv[1], argv[2])
 
 
 if __name__ == "__main__":
